+++ i18n/giswater_en_US.ts
@@ -10700,6 +10700,22 @@ or they were created by another user:</translation>
 			<translation>groupBox_3</translation>
 		</message>
 		<message>
+			<source>groupBox_4</source>
+			<translation>Nodes</translation>
+		</message>
+		<message>
+			<source>tooltip_groupBox_4</source>
+			<translation>groupBox_4</translation>
+		</message>
+		<message>
+			<source>groupBox_5</source>
+			<translation>Extra filters</translation>
+		</message>
+		<message>
+			<source>tooltip_groupBox_5</source>
+			<translation>groupBox_5</translation>
+		</message>
+		<message>
 			<source>tab_connect</source>
 			<translation>Link</translation>
 		</message>